fix: prevent app reload when saving anything within app/components/UI/Trending
